jetson-ml-agent/requirements.txt
flask==3.0.0
requests==2.32.3
numpy>=1.21.2,<2.0
tensorflow==2.13.1
onnxruntime>=1.16,<1.19
pyserial>=3.5
pyusb==1.2.1
